moved helper functions to utils, removed tests that can't succeed for now, renamed tests to follow naming of their respective classes
