+++ test/utils/test_AudioBuffer.cpp
@@ -2,7 +2,7 @@
 #include <anira/anira.h>
 
 using namespace anira;
-TEST(Audiobuffer, SimpleWrite){
+TEST(AudioBuffer, SimpleWrite){
     AudioBufferF buffer = AudioBufferF(1,10);
     for (size_t i = 0; i < buffer.get_num_samples(); i++){
         EXPECT_FLOAT_EQ(0.f, buffer.get_sample(0,i));
@@ -21,7 +21,7 @@ TEST(Audiobuffer, SimpleWrite){
     }
 }
 
-TEST(Audiobuffer, BlockSwap){
+TEST(AudioBuffer, BlockSwap){
     int block_size = 10;
 
     MemoryBlock<int> block;
@@ -60,7 +60,7 @@ TEST(Audiobuffer, BlockSwap){
     }
 }
 
-TEST(Audiobuffer, BufferSwap){
+TEST(AudioBuffer, BufferSwap){
     int block_size = 10;
 
     anira::AudioBuffer<int> buffer1(1, block_size);
@@ -96,7 +96,7 @@ TEST(Audiobuffer, BufferSwap){
         ASSERT_EQ(buffer2.get_sample(0,i), i);
     }
 }
-TEST(Audiobuffer, InvalidSizeSwap){
+TEST(AudioBuffer, InvalidSizeSwap){
     anira::AudioBuffer<int> buffer1(1, 5);
     anira::AudioBuffer<int> buffer2(1, 6);
     int* buffer1_ptr = buffer1.data();
@@ -113,7 +113,7 @@ TEST(Audiobuffer, InvalidSizeSwap){
     ASSERT_EQ(output, std::string("Cannot swap data, buffers have different number of channels or sizes!\n"));
 }
 
-TEST(Audiobuffer, InvalidChannelsSwap){
+TEST(AudioBuffer, InvalidChannelsSwap){
     anira::AudioBuffer<int> buffer1(2, 5);
     anira::AudioBuffer<int> buffer2(1, 5);
     int* buffer1_ptr = buffer1.data();
@@ -130,7 +130,7 @@ TEST(Audiobuffer, InvalidChannelsSwap){
     ASSERT_EQ(output, std::string("Cannot swap data, buffers have different number of channels or sizes!\n"));
 }
 
-// TEST(Audiobuffer, BlockOfBlocks){
+// TEST(AudioBuffer, BlockOfBlocks){
 //     int block_size = 10;
 //     MemoryBlock<std::atomic<int>> block_of_atomics(block_size);
 //     for (int i = 0; i < block_size; i++) {
